Move genres component to book list dir
